chore: Update 3D map export dialog component tests
